Extract useCompiledStyles to dedicated file
