+++ repos/re-theme/src/helpers/index.js
@@ -1,1 +1,2 @@
 export * from './getTheme'
+export * from './compileStyles'
