+++ repos/re-theme/src/hooks/index.js
@@ -5,3 +5,4 @@ export * from './useThemeHover'
 export * from './useThemeActive'
 export * from './useThemeFocus'
 export * from './useStyle'
+export * from './useCompiledStyles'
